app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Hossein Sokhandan <iamsokhandan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.72.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.1"}
aerich = "^0.6.2"
autopep8 = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
